backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
pdfplumber
python-dotenv
pydantic
